requirements.txt
jinja2
pydantic
# pyyaml
pytoml
textual
typedmodel